// src/interfaces/course.interface.ts
export interface Course {
    id: number;
    title: string;
    image_url: string;
    image_id: string;
    category: string;
    price: number;
    creator: string;

  }

  export interface Cou {
    title: string;
    image_url: string;
    image_id: string;
    category: string;
    price: number;
    creator: string;
  }